+++ src/preserve_env_list.h
@@ -1,0 +1,35 @@
+/*
+    libfakechroot -- fake chroot environment
+    Copyright (c) 2013 Piotr Roszatycki <dexter@debian.org>
+
+    This library is free software; you can redistribute it and/or
+    modify it under the terms of the GNU Lesser General Public
+    License as published by the Free Software Foundation; either
+    version 2.1 of the License, or (at your option) any later version.
+
+    This library is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+    Lesser General Public License for more details.
+
+    You should have received a copy of the GNU Lesser General Public
+    License along with this library; if not, write to the Free Software
+    Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307  USA
+*/
+
+
+/* List of environment variables to preserve on clearenv() */
+char *preserve_env_list[] = {
+    "FAKECHROOT_BASE",
+    "FAKECHROOT_CMD_SUBST",
+    "FAKECHROOT_DEBUG",
+    "FAKECHROOT_DETECT",
+    "FAKECHROOT_ELFLOADER",
+    "FAKECHROOT_ELFLOADER_OPT_ARGV0",
+    "FAKECHROOT_EXCLUDE_PATH",
+    "FAKECHROOT_VERSION",
+    "FAKEROOTKEY",
+    "FAKED_MODE",
+    "LD_LIBRARY_PATH",
+    "LD_PRELOAD"
+};